requirements.txt
PyMuPDF
Flask
requests
Pillow
